orchestrator_service/requirements.txt
fastapi
uvicorn
pydantic
httpx
tenacity
structlog
asyncpg
redis
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
python-dotenv
prometheus-client
requests
python-dateutil
python-socketio
aiofiles
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
pyjwt
email-validator